requirements-dev.txt
django-nose>=1.2
nose>=1.3,<2
nosexcover>=1.0.8
ipdb